refactor: separate customizers and utilities
This PR also moves the source files to `src/` folder.
