+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.2.14",
   "description": "",
   "repository": "arackaf/customize-cra",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
     "test": "jest"
   },
