+++ src/customizers.js
@@ -1,4 +1,4 @@
-const flow = require("lodash.flow");
+const { getBabelLoader } = require("./utilities");
 
 const addBundleVisualizer = (options = {}, behindFlag = false) => config => {
   const BundleAnalyzerPlugin = require("webpack-bundle-analyzer")
@@ -20,29 +20,6 @@ const addBundleVisualizer = (options = {}, behindFlag = false) => config => {
     );
   }
   return config;
-};
-
-const getBabelLoader = config => {
-  const babelLoaderFilter = rule =>
-    rule.loader &&
-    rule.loader.includes("babel") &&
-    rule.options &&
-    rule.options.plugins;
-
-  // First, try to find the babel loader inside the oneOf array.
-  // This is where we can find it when working with react-scripts@2.0.3.
-  let loaders = config.module.rules.find(rule => Array.isArray(rule.oneOf))
-    .oneOf;
-
-  let babelLoader = loaders.find(babelLoaderFilter);
-
-  // If the loader was not found, try to find it inside of the "use" array, within the rules.
-  // This should work when dealing with react-scripts@2.0.0.next.* versions.
-  if (!babelLoader) {
-    loaders = loaders.reduce((ldrs, rule) => ldrs.concat(rule.use || []), []);
-    babelLoader = loaders.find(babelLoaderFilter);
-  }
-  return babelLoader;
 };
 
 const addBabelPlugin = plugin => config => {
@@ -151,8 +128,6 @@ const babelInclude = include => config => {
   return config;
 };
 
-const override = (...plugins) => flow(...plugins.filter(f => f));
-
 const addBabelPlugins = (...plugins) => plugins.map(p => addBabelPlugin(p));
 
 const addBabelPresets = (...plugins) => plugins.map(p => addBabelPreset(p));
@@ -251,17 +226,6 @@ const addLessLoader = (loaderOptions = {}) => config => {
   return config;
 };
 
-// Use this helper to override the webpack dev server settings
-//  it works just like the `override` utility
-const overrideDevServer = (...plugins) => configFunction => (
-  proxy,
-  allowedHost
-) => {
-  const config = configFunction(proxy, allowedHost);
-  const updatedConfig = override(...plugins)(config);
-  return updatedConfig;
-};
-
 // to be used inside `overrideDevServer`, makes CRA watch all the folders
 // included `node_modules`, useful when you are working with linked packages
 // usage: `yarn start --watch-all`
@@ -337,7 +301,6 @@ const addTslintLoader = options => config => {
 };
 
 module.exports = {
-  override,
   addBundleVisualizer,
   addBabelPlugin,
   addDecoratorsLegacy,
@@ -353,14 +316,12 @@ module.exports = {
   fixBabelImports,
   useBabelRc,
   addLessLoader,
-  overrideDevServer,
   watchAll,
   babelInclude,
   addBabelPreset,
   addBabelPresets,
   disableChunk,
   addPostcssPlugins,
-  getBabelLoader,
   removeModuleScopePlugin,
   addTslintLoader
 };
