+++ src/index.js
@@ -1,0 +1,7 @@
+const customizers = require("./customizers");
+const utilities = require("./utilities");
+
+module.exports = {
+  ...customizers,
+  ...utilities
+};
